+++ Sandswept/Items/Whites/FracturedTimepiece.cs
@@ -9,9 +9,9 @@ namespace Sandswept.Items.Whites
 
         public override string ItemLangTokenName => "FRACTURED_TIMEPIECE";
 
-        public override string ItemPickupDesc => "Using your utility skill heals you and reduces special cooldown.";
+        public override string ItemPickupDesc => "Using your Utility skill heals you and reduces Special skill cooldown.";
 
-        public override string ItemFullDescription => ("Upon using your $suutility skill$se, $shheal$se for $sh" + d(basePercentHealing) + "$se $ss(+" + d(stackPercentHealing) + " per stack)$se of your $shmaximum health$se and $sureduce special skill cooldown$se by $su" + d(specialCooldownReduction) + "$se.").AutoFormat();
+        public override string ItemFullDescription => $"Upon using your $suUtility skill$se, $shheal$se for $sh{basePercentHealing * 100f}%$se $ss(+{stackPercentHealing * 100f}% per stack)$se of your $shmaximum health$se and $sureduce Special skill cooldown$se by $su{specialCooldownReduction * 100f}%$se.".AutoFormat();
 
         public override string ItemLore => "<style=cMono>Order: Timepiece\r\nTracking Number: 864*******\r\nEstimated Delivery: 02/23/2054\r\nShipping Method: Priority\r\nShipping Address: Hall of the Revered, Mars\r\nShipping Details:\r\n\r\n</style>Our team uncovered this in the desert where the old Hall was, before it burned down. I guess someone really wanted to protect it from the Purge, since it was carefully wrapped and boxed where we found it. You can probably glean more knowledge from it than we can, and it was the Hall's to begin with in any case.\r\n\r\nThe box had a note in it from the Time Keeper of the era, too, which I've included in the package. Nobody here can read the old language, so hopefully you can make some sense of it.\r\n";